Rehash description to reflect project scope and outcomes
Address the vague and uninformative description which was appended
to the README.md file on creation of repo.

Articulate the structure and design of the project, scope of knowledge
being examined, as well as the desired learning outcomes.

This initial description serves as a benchmark against which a final
assessment can be conducted to measure any differences between our
expected and realized learning outcomes.